api/requirements.txt
python-dotenv~=1.0.0
uvicorn~=0.22.0
redis~=4.5.5
fastapi~=0.96.0
starlette~=0.27.0
google-cloud-storage